requirement.txt
python-dotenv==1.0.0
langchain==0.3.8
langchain-openai==0.3.2
langchain-community==0.3.4
openai==1.61.0
numpy==1.23.5
gensim==4.3.1
nltk==3.8.1
wikipedia==1.4.0
scikit-learn==1.2.2
